requirements.txt
requests>=2.31.0
pytest>=7.4.0
python-dotenv>=1.0.1
Pillow>=10.0.0
opencv-python-headless>=4.10.0


